% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printf.R
\name{d3s_printf}
\alias{d3s_printf}
\title{C-like printf}
\usage{
d3s_printf(fmt, ...)
}
\arguments{
\item{fmt}{printf-style format}

\item{...}{Arguments to be formatted.}
}
\description{
Behaves exactly like printf from C: no automatic line endings etc.
}
