% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger_banner}
\alias{logger_banner}
\title{Logs a banner message.}
\usage{
logger_banner(fmt, ...)
}
\arguments{
\item{fmt}{sprintf-like formatting string.}

\item{...}{Values to be passed to \code{fmt}.}
}
\description{
Useful for delimiting big parts of your long-running scripts.
}
