% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger_crazy}
\alias{logger_crazy}
\title{Logs a detailed debugging message.}
\usage{
logger_crazy(fmt, ...)
}
\arguments{
\item{fmt}{sprintf-like formatting string.}

\item{...}{Values to be passed to \code{fmt}.}
}
\description{
Logs a detailed debugging message.
}
