% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger_debug}
\alias{logger_debug}
\title{Logs a debugging message.}
\usage{
logger_debug(fmt, ...)
}
\arguments{
\item{fmt}{sprintf-like formatting string.}

\item{...}{Values to be passed to \code{fmt}.}
}
\description{
Logs a debugging message.
}
