% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger_info}
\alias{logger_info}
\title{Logs an information message.}
\usage{
logger_info(fmt, ...)
}
\arguments{
\item{fmt}{sprintf-like formatting string.}

\item{...}{Values to be passed to \code{fmt}.}
}
\description{
Logs an information message.
}
