% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger_level}
\alias{logger_level}
\title{Set logging level.}
\usage{
logger_level(new_level)
}
\arguments{
\item{new_level}{One of \code{warn}, \code{info}, \code{debug} or \code{crazy}.}

\item{...}{Arguments to be formatted.}
}
\description{
Messages with higher (more detailed) level will not be printed.
}
