% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger_progress_init}
\alias{logger_progress_init}
\title{Initializes logger-based progress messages.}
\usage{
logger_progress_init(msg, task_count, steps = 0.1)
}
\arguments{
\item{task_count}{How many tasks to expect.}

\item{steps}{How many tasks has to finish before message is printed.}

\item{fmt}{sprintf-like formatting string that will receive percentage.}
}
\description{
Initializes logger-based progress messages.
}
