% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger_warn}
\alias{logger_warn}
\title{Logs a warning message.}
\usage{
logger_warn(fmt, ...)
}
\arguments{
\item{fmt}{sprintf-like formatting string.}

\item{...}{Values to be passed to \code{fmt}.}
}
\description{
Logs a warning message.
}
