% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_jobs.R
\name{simple_jobs_collect}
\alias{simple_jobs_collect}
\title{Waits for completion of all simple-jobs.}
\usage{
simple_jobs_collect()
}
\value{
Task results as list, indexed by task name.
}
\description{
Waits for completion of all simple-jobs.
}
