% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_jobs.R
\name{simple_jobs_init}
\alias{simple_jobs_init}
\title{Initializes simple-jobs.}
\usage{
simple_jobs_init(max_jobs = NULL)
}
\arguments{
\item{max_jobs}{Maximum number of concurrent jobs.}
}
\description{
Initializes simple-jobs.
}
