% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_jobs.R
\name{simple_jobs_submit}
\alias{simple_jobs_submit}
\title{Submits new task to simple-jobs.}
\usage{
simple_jobs_submit(expr, name = NULL)
}
\arguments{
\item{expr}{Expression to evaluate asynchronously in a new thread (process).}

\item{name}{Custom name of the new task.}
}
\value{
Job name
}
\description{
Submits new task to simple-jobs.
}
